UPSTREAM: erofs-utils: lib: Fix calculation of minextblks when working with sparse files
When we work with sparse files (files with holes), we need to consider
when the contiguous data block starts after each hole to correctly calculate
minextblks so we can merge consecutive chunks later.
Now that we need to recalculate minextblks multiple places, put the logic
in helper function for avoiding repetition and easier reading.

Fixes: 7b46f7a0160a ("erofs-utils: lib: merge consecutive chunks if possible")
Signed-off-by: Sandeep Dhavale <dhavale@google.com>
Reviewed-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Link: https://lore.kernel.org/r/20240403070700.1716252-1-dhavale@google.com
Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Bug: 331693570
Test: Th
Test: Create erofs image with sparse files, mount, verify integrity
using md5sum
Link:
https://lore.kernel.org/linux-erofs/20240403062357.1705807-1-dhavale@google.com/
Change-Id: Ib5b89c44064bb3feb44e00fe92a3fecee578ed78
Signed-off-by: Sandeep Dhavale <dhavale@google.com>